requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==42.0.4
discord==2.1.0
discord.py==2.1.0
frozendict==2.3.4
frozenlist==1.3.3
html5lib==1.1
idna==3.7
lxml==4.9.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.1
pandas==1.5.2
polygon-api-client==1.7.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.32.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.18
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
yfinance==0.2.9
